Add ProcessHelper utility class
