+++ src/OmniSharp.Abstractions/Utilities/PlatformHelper.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.IO;
 
 namespace OmniSharp.Utilities
@@ -18,7 +17,7 @@ namespace OmniSharp.Utilities
         private static string FindMonoPath()
         {
             // To locate Mono on unix, we use the 'which' command (https://en.wikipedia.org/wiki/Which_(Unix))
-            var monoFilePath = LaunchProcessAndCaptureOutput("which", "mono");
+            var monoFilePath = ProcessHelper.RunAndCaptureOutput("which", "mono");
 
             if (string.IsNullOrEmpty(monoFilePath))
             {
@@ -88,7 +87,7 @@ namespace OmniSharp.Utilities
                 // the final path.
 
                 var originalPath = paths[paths.Count - 1];
-                var newPath = LaunchProcessAndCaptureOutput("readlink", $"{originalPath}");
+                var newPath = ProcessHelper.RunAndCaptureOutput("readlink", $"{originalPath}");
 
                 if (string.IsNullOrEmpty(newPath) ||
                     string.CompareOrdinal(originalPath, newPath) == 0)
@@ -145,34 +144,7 @@ namespace OmniSharp.Utilities
         private static string CanonicalizeDirectory(string directoryName)
         {
             // Use "pwd -P" to get the directory name with all symbolic links on Unix.
-            return LaunchProcessAndCaptureOutput("pwd", "-P", directoryName);
-        }
-
-        private static string LaunchProcessAndCaptureOutput(string command, string args, string workingDirectory = null)
-        {
-            var startInfo = new ProcessStartInfo(command, args);
-            startInfo.RedirectStandardOutput = true;
-            startInfo.RedirectStandardError = true;
-            startInfo.CreateNoWindow = true;
-            startInfo.UseShellExecute = false;
-            startInfo.WorkingDirectory = workingDirectory ?? string.Empty;
-
-            var process = new Process();
-            process.StartInfo = startInfo;
-
-            try
-            {
-                process.Start();
-                var output = process.StandardOutput.ReadToEnd();
-                process.WaitForExit();
-
-                return output.Trim();
-            }
-            catch
-            {
-                Console.WriteLine($"Failed to launch '{command}' with args, '{args}'");
-                return null;
-            }
+            return ProcessHelper.RunAndCaptureOutput("pwd", "-P", directoryName);
         }
     }
 }
